examples/flappy/requirements.txt
# PyGame Learning Environment
pygame==1.9.5
gym_ple==0.3
